requirements.txt
pandas
xlrd
pyyaml
torch
transformers
datasets
sacrebleu
bert-score
unbabel-comet
accelerate>=0.26.0
